workflow/pyproject.toml
[project]
name = "workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "guppy3>=3.1.5",
    "humanize>=4.13.0",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "obstore>=0.7.1",
    "pillow>=11.3.0",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.1",
    "rasterio>=1.4.3",
    "tilebox-datasets>=0.39.0",
    "tilebox-storage>=0.39.0",
    "tilebox-workflows>=0.39.0",
    "xarray>=2025.7.1",
    "zarr>=3.1.0",
]

[dependency-groups]
dev = ["pyright>=1.1.404", "ruff>=0.12.7"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
packages = ["fpar_to_vci"]

[tool.hatch.build.targets.wheel]
packages = ["fpar_to_vci"]


[tool.ruff]
# settings applying to all ruff subcommands, such as `ruff check` (lint) and `ruff format`
line-length = 120
target-version = "py310"
exclude = ["*/.venv/*"]

[tool.ruff.lint]
select = ["ALL"]
# all rules: https://beta.ruff.rs/docs/rules
ignore = [
    # some ruff checking modules don't make much sense as a whole
    "D",   # pydocstyle: pretty strict anyways
    "FBT", # flake8-boolean-trap: boolean arguments can make sense
    "COM", # flake8-commas: formatter takes care of this
    "DTZ", # flake8-datetimez: utc datetimes are useful
    "DJ",  # flake8-django: not needed
    "EM",  # flake8-errmsg: str directly in Exception constructor is accetable
    "TCH", # flake8-type-checking: type checking blocks are weird
    # specific rules
    "ANN401",  # any-type: allow Any in *args and **kwargs
    "S101",    # assert: allow usage of assert
    "B008",    # function-call-argument-default: some default argument values make sense to be function calls
    "G004",    # logging-f-string: allow usage of f-strings in logging calls
    "PLR2004", # magic-value-comparison: sometimes comparison with constants (e.g. 0) makes sense
    "TRY003",  # raise-vanilla-args: exceptions like this make sense in python
    "TRY400",  # error-instead-of-exception: logger.error is ok with loguru
    # disabled because of formatter
    "E501",   # line-too-long -> formatter takes care of this
    "ISC001", # single-line-implicit-str-concatenation -> formatter takes care of this
    "Q",      # flake8-quotes -> formatter takes care of this
    "W191",   # tab-indentation -> formatter takes care of this
]

unfixable = [
    "F841", # unused-variable -> don't remove them automatically
    "ERA",  # eradicate -> (commented out code), don't remove it automatically
]

[tool.ruff.lint.isort]
known-first-party = ["tilebox", "_tilebox"]

[tool.ruff.lint.per-file-ignores]
"*/tests/*" = ["INP001", "SLF001"]

[tool.pyright]
exclude = ["**/.ipynb_checkpoints", "**/__pycache__", ".venv"]


# pyright needs to have all the dependencies installed to be able to type check
# we can make sure of this by telling it to use the uv venv
venvPath = "."
venv = ".venv"


reportPrivateImportUsage = false
